add fuctions open camera, edit and delete
