+++ android/build.gradle
@@ -48,6 +48,7 @@ allprojects {
             }
         }
         google()
+        maven { url 'https://maven.google.com' }
         maven { url 'https://www.jitpack.io' }
     }
 }
