+++ package.json
@@ -25,7 +25,7 @@
     "react-native-file-picker": "^0.0.21",
     "react-native-fs": "^2.20.0",
     "react-native-gesture-handler": "^2.4.2",
-    "react-native-image-crop-tools": "^1.6.4",
+    "react-native-image-crop-picker": "^0.35.3",
     "react-native-image-picker": "^4.10.3",
     "react-native-image-zoom-viewer": "^3.0.1",
     "react-native-modal": "^13.0.1",
