--- src/components/EditImage.js
@@ -1,13 +1,0 @@
-import React from 'react'
-import { View,Text } from 'react-native'
-//libtrtia 
-
-const EditImage=({uri})=> {
-  return (
-    <View>
-        <Text>Hola Mundo</Text>
-    </View>
-  )
-}
-
-export default EditImage
